% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrRandomForest.R
\docType{data}
\name{LearnerRegrRandomForest}
\alias{LearnerRegrRandomForest}
\alias{mlr_learners_regr.randomForest}
\title{Regression Random Forest Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrRandomForest
}
\description{
A \link{LearnerRegr} for a regression random forest implemented in randomForest::randomForest()] in package \CRANpkg{randomForest}.
}
\references{
Breiman, L. (2001).
Random Forests
Machine Learning
\url{https://doi.org/10.1023/A:1010933404324}
}
\keyword{datasets}
