% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.randomForest-package}
\alias{mlr3learners.randomForest}
\alias{mlr3learners.randomForest-package}
\title{mlr3learners.randomForest: Random forest for mlr3 (randomForest package)}
\description{
Wrapper for a classification and regression learner from the
  randomForest package (CRAN) to be used in mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners}
  \item \url{https://github.com/mlr3learners/mlr3learners.earth}
}

}
\author{
\strong{Maintainer}: Philipp Kopper \email{kopperphilipp@gmail.com}

Authors:
\itemize{
  \item Quay Au \email{quayau@gmail.com}
}

}
