% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.randomforest-package}
\alias{mlr3learners.randomforest}
\alias{mlr3learners.randomforest-package}
\title{mlr3learners.randomforest: Random forest for mlr3 (randomForest package)}
\description{
Wrapper for a classification and regression learner
    from the randomForest package (CRAN) to be used in mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.randomforest}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

Authors:
\itemize{
  \item Philipp Kopper \email{kopperphilipp@gmail.com}
  \item Quay Au \email{quayau@gmail.com}
}

}
